requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<3.17
djangorestframework-gis>=1.2
django-filter>=25.2
django-leaflet>=0.33
django-colorfield>=0.14
django-smart-selects @ git+https://github.com/jazzband/django-smart-selects.git@1f3ac7c08a31d2e8cc792f1de484a027a25eef1f
pillow>=12
psycopg[binary]>=3.2
gunicorn>=23
